move isActivityBuilder to api package + added typedocs
