+++ packages/api/src/activities/index.ts
@@ -1,13 +1,12 @@
 import { EventActivity } from './event';
 import { InvokeActivity } from './invoke';
 import { TraceActivity } from './trace';
-import { TypingActivity, TypingActivityBuilder } from './typing';
+import { TypingActivity } from './typing';
 import { HandoffActivity } from './handoff';
 import { ConversationActivity } from './conversation';
-import { MessageActivity, MessageDeleteActivityBuilder, MessageReactionActivityBuilder, MessageSendActivityBuilder, MessageUpdateActivityBuilder } from './message';
+import { MessageActivity } from './message';
 import { CommandActivity } from './command';
 import { InstallUpdateActivity } from './install-update';
-import { ActivityBaseBuilder } from './base';
 
 export type Activity =
   | MessageActivity
@@ -20,21 +19,6 @@ export type Activity =
   | CommandActivity
   | InstallUpdateActivity;
 
-export type ActivityBuilder = 
-  | MessageDeleteActivityBuilder
-  | MessageReactionActivityBuilder
-  | MessageSendActivityBuilder
-  | MessageUpdateActivityBuilder
-  | TypingActivityBuilder;
-
-/**
- * @hidden
- * @internal
- */
-export function isActivityBuilder(value: any): value is ActivityBuilder {
-  return value instanceof ActivityBaseBuilder;
-}
-
 export * from './message';
 export * from './event';
 export * from './invoke';
@@ -44,3 +28,5 @@ export * from './handoff';
 export * from './conversation';
 export * from './command';
 export * from './install-update';
+export * from "./utils";
+export * from "./builder";
